<chore>[sdk]: merge ErrorCodeList to ErrorCode in sdk
Deserializing ErrorCode using class JsonObjectUtil
will lose the information of "ErrorCodeList.cases",
so the "causes" field will be directly merged into ErrorCode.
