--- sdk/src/main/java/org/zstack/sdk/ErrorCodeList.java
@@ -1,15 +1,0 @@
-package org.zstack.sdk;
-
-
-
-public class ErrorCodeList extends org.zstack.sdk.ErrorCode {
-
-    public java.util.List causes;
-    public void setCauses(java.util.List causes) {
-        this.causes = causes;
-    }
-    public java.util.List getCauses() {
-        return this.causes;
-    }
-
-}
